Renameed DocumentType to DocumentTypes
