+++ src/main/java/uk/gov/hmcts/reform/adoption/document/DocumentUtil.java
@@ -16,7 +16,7 @@ public final class DocumentUtil {
             documentInfo.getBinaryUrl());
     }
 
-    public static AdoptionDocument adoptionDocumentFrom(final DocumentInfo documentInfo, final DocumentType documentType) {
+    public static AdoptionDocument adoptionDocumentFrom(final DocumentInfo documentInfo, final DocumentTypes documentType) {
         return AdoptionDocument
             .builder()
             .documentLink(documentFrom(documentInfo))
