+++ src/test/java/uk/gov/hmcts/reform/adoption/testutil/TestDataHelper.java
@@ -2,23 +2,23 @@ package uk.gov.hmcts.reform.adoption.testutil;
 
 import uk.gov.hmcts.ccd.sdk.type.Document;
 import uk.gov.hmcts.ccd.sdk.type.ListValue;
-import uk.gov.hmcts.reform.adoption.document.DocumentType;
+import uk.gov.hmcts.reform.adoption.document.DocumentTypes;
 import uk.gov.hmcts.reform.adoption.document.model.AdoptionDocument;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
 
-import static uk.gov.hmcts.reform.adoption.document.DocumentType.APPLICATION;
+import static uk.gov.hmcts.reform.adoption.document.DocumentTypes.APPLICATION;
 
 public class TestDataHelper {
 
     public static final LocalDateTime LOCAL_DATE_TIME = LocalDateTime.of(2021, 4, 28, 1, 0);
 
-    public static ListValue<AdoptionDocument> documentWithType(final DocumentType documentType) {
+    public static ListValue<AdoptionDocument> documentWithType(final DocumentTypes documentType) {
         return documentWithType(documentType, UUID.randomUUID().toString());
     }
 
-    public static ListValue<AdoptionDocument> documentWithType(final DocumentType documentType,
+    public static ListValue<AdoptionDocument> documentWithType(final DocumentTypes documentType,
                                                               final String documentId) {
         String documentUrl = "http://localhost:8080/" + documentId;
